moved server to index.ts because I don't want to configure.
